backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
minio==7.2.3
celery[redis]==5.3.6
redis[hiredis]>=5.0.0
python-docx==1.1.0
openpyxl==3.1.2
pdfplumber==0.10.3
python-magic==0.4.27
aiofiles==23.2.1
websockets==12.0
httpx==0.27.0
pytest==8.0.1
pytest-asyncio==0.23.5
